Docker/Hacker-News-Api/requirements.txt
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==24.2.0
cattrs==23.2.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.103.1
fastapi-cache2==0.1.8
fastapi-redis-cache==0.2.5
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
pendulum==3.0.0
platformdirs==4.2.2
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
requests-cache==1.2.1
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
time-machine==2.15.0
typing_extensions==4.12.2
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.2
uvicorn==0.23.2
yarl==1.9.4
